move traverse_pre_order to convert.rs
